requirements.txt
absl-py==1.3.0
async-generator==1.10
asyncio==3.4.3
attrs==22.1.0
bagit==1.8.1
beautifulsoup4==4.11.1
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
contourpy==1.0.6
cycler==0.11.0
distlib==0.3.6
docopt==0.6.2
exceptiongroup==1.0.0rc9
filelock==3.8.0
Flask==2.2.2
flatbuffers==22.11.23
fonttools==4.38.0
fpdf==1.7.2
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
identify==2.5.9
idna==2.10
importlib-metadata==5.1.0
imutils==0.5.4
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
Js2Py==0.71
kiwisolver==1.4.4
libclang==15.0.6.1
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.2
mediapipe==0.9.0
MouseInfo==0.1.3
mypy-extensions==0.4.3
nodeenv==1.7.0
numpy==1.23.5
opencv-contrib-python==4.6.0.66
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pathspec==0.10.2
Pillow==9.3.0
pipwin==0.5.2
platformdirs==2.5.4
pre-commit==2.20.0
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
PyAutoGUI==0.9.53
pycaw==20220416
pycparser==2.21
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
pypng==0.20220715.0
PyPrind==2.11.3
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
pySmartDL==1.3.4
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pywhatkit==5.4
pywin32==304
PyYAML==6.0
pyzbar==0.1.9
requests==2.28.1
rfc3986==1.5.0
robotframework==6.0.2
robotframework-ocrlibrary==2.0.0
rsa==4.9
scipy==1.9.3
selenium==4.5.0
semantic-version==2.10.0
setuptools-rust==1.5.2
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.2.3
toml==0.10.2
tomli==2.0.1
torch==1.13.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
virtualenv==20.17.0
Werkzeug==2.2.2
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.2.0
zipp==3.11.0
